requirements.txt
streamlit
mistralai
google-genai
re
python-dotenv
streamlit-mermaid
langchain-google-genai
langchain-core
faiss-cpu